requirements.txt
pandas
gymnasium
pytorch-lightning 
-e .